+++ screens/FavoritesScreen.js
@@ -1,21 +1,14 @@
 import React from 'react';
-import {Text, View, StyleSheet, Image} from 'react-native';
+
+import MealList from '../components/MealList';
+import { MEALS } from '../data/dummy-data';
 
 const FavoriteScreen = () => {
-  return (
-    <View style={styles.container}>
-      <Image style={{width: 300, height: 300}} source={require('../models/images/meals.png')} />
-      <Text>FavoriteScreen</Text>
-    </View>
-  );
+
+  const favMeals = MEALS.filter(meal => meal.id === 'm2 ' || meal.id === 'm1')
+
+  return <MealList listData={favMeals} />
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
 
 export default FavoriteScreen;
